+++ spec/features/admin_claim_tasks_update_with_dqt_api_spec.rb
@@ -87,9 +87,12 @@ RSpec.feature "Admin claim tasks update with DQT API" do
           programme_end_date: "2021-07-04T00:00:00Z",
           programme_type: "Overseas Trained Teacher Programme",
           result: "Pass",
-          subject1: data.dig(:itt_subject_codes, 0).to_s,
-          subject2: data.dig(:itt_subject_codes, 1).to_s,
-          subject3: data.dig(:itt_subject_codes, 2).to_s,
+          subject1: data.dig(:itt_subjects, 0).to_s,
+          subject1_code: data.dig(:itt_subject_codes, 0).to_s,
+          subject2: data.dig(:itt_subjects, 1).to_s,
+          subject2_code: data.dig(:itt_subject_codes, 1).to_s,
+          subject3: data.dig(:itt_subjects, 2).to_s,
+          subject3_codes: data.dig(:itt_subject_codes, 2).to_s,
           qualification: (data[:qualification_name] || "BA").to_s,
           state: 0,
           state_name: "Active"
@@ -627,6 +630,7 @@ RSpec.feature "Admin claim tasks update with DQT API" do
               9,
               1
             ),
+            itt_subjects: ["mathematics"],
             itt_subject_codes: [EarlyCareerPayments::DqtRecord::ELIGIBLE_HECOS_CODES[:mathematics].first]
           }
         end
@@ -667,6 +671,7 @@ RSpec.feature "Admin claim tasks update with DQT API" do
                 9,
                 1
               ),
+              itt_subjects: ["applied mathematics", "maths and statistics", :maths],
               itt_subject_codes: [EarlyCareerPayments::DqtRecord::ELIGIBLE_HECOS_CODES[:mathematics].first]
             }
           end
@@ -708,6 +713,7 @@ RSpec.feature "Admin claim tasks update with DQT API" do
                 9,
                 1
               ),
+              itt_subjects: [],
               itt_subject_codes: ["NoCode"]
             }
           end
@@ -788,6 +794,7 @@ RSpec.feature "Admin claim tasks update with DQT API" do
               9,
               1
             ),
+            itt_subjects: [],
             itt_subject_codes: ["NoCode"]
           }
         end
@@ -1326,6 +1333,7 @@ RSpec.feature "Admin claim tasks update with DQT API" do
               9,
               1
             ),
+            itt_subjects: ["applied mathematics", "maths and statistics", "mathematics"],
             itt_subject_codes: [MathsAndPhysics::DqtRecord::ELIGIBLE_MATHS_HECOS_CODES.first]
           }
         end
@@ -1366,6 +1374,7 @@ RSpec.feature "Admin claim tasks update with DQT API" do
                 9,
                 1
               ),
+              itt_subjects: ["applied mathematics", "maths and statistics", "mathematics"],
               itt_subject_codes: [MathsAndPhysics::DqtRecord::ELIGIBLE_MATHS_HECOS_CODES.first]
             }
           end
@@ -1407,6 +1416,7 @@ RSpec.feature "Admin claim tasks update with DQT API" do
                 9,
                 1
               ),
+              itt_subjects: [],
               itt_subject_codes: ["NoCode"]
             }
           end
@@ -1486,6 +1496,7 @@ RSpec.feature "Admin claim tasks update with DQT API" do
               9,
               1
             ),
+            itt_subjects: ["applied mathematics", "maths and statistics", "maths"],
             itt_subject_codes: ["NoCode"]
           }
         end
@@ -2024,6 +2035,7 @@ RSpec.feature "Admin claim tasks update with DQT API" do
               9,
               1
             ),
+            itt_subjects: ["SUBJECT"],
             itt_subject_codes: ["CODE"]
           }
         end
@@ -2064,6 +2076,7 @@ RSpec.feature "Admin claim tasks update with DQT API" do
                 9,
                 1
               ),
+              itt_subjects: ["SUBJECT"],
               itt_subject_codes: ["CODE"]
             }
           end
@@ -2145,6 +2158,7 @@ RSpec.feature "Admin claim tasks update with DQT API" do
               9,
               1
             ),
+            itt_subjects: ["Subject"],
             itt_subject_codes: ["NoCode"]
           }
         end