+++ spec/models/early_career_payments/dqt_record_spec.rb
@@ -19,6 +19,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
     OpenStruct.new(
       {
         degree_codes: record_degree_codes,
+        itt_subjects: record_itt_subjects,
         itt_subject_codes: record_itt_subject_codes,
         itt_start_date: record_itt_date,
         qts_award_date: record_qts_date,
@@ -31,9 +32,11 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
     subject(:eligible?) { dqt_record.eligible? }
 
     [
+      # start of mathematics for 2021
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["G100"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2018"),
@@ -42,6 +45,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["G100"],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2017"),
@@ -50,6 +54,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: ["G100"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2018"),
@@ -58,6 +63,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: ["G100"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2017"),
@@ -66,6 +72,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["100403"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2018"),
@@ -74,6 +81,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["100403"],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2017"),
@@ -82,6 +90,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: ["100403"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2018"),
@@ -90,14 +99,19 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: ["100403"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2017"),
         record_qualification_name: nil
       },
+      # end of mathematics for 2021
+
+      # start of mathematics for 2022
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["G100"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2019"),
@@ -106,6 +120,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["G100"],
         record_itt_date: Date.parse("1/9/2019"),
         record_qts_date: Date.parse("1/9/2017"),
@@ -114,6 +129,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: ["G100"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2019"),
@@ -122,6 +138,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: ["G100"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2019"),
         record_qts_date: Date.parse("1/9/2017"),
@@ -130,6 +147,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["100403"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2019"),
@@ -138,6 +156,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["100403"],
         record_itt_date: Date.parse("1/9/2019"),
         record_qts_date: Date.parse("1/9/2017"),
@@ -146,6 +165,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: ["100403"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2019"),
@@ -154,6 +174,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: ["100403"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2019"),
         record_qts_date: Date.parse("1/9/2017"),
@@ -162,6 +183,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["G100"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -170,6 +192,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["G100"],
         record_itt_date: Date.parse("1/9/2020"),
         record_qts_date: Date.parse("1/9/2017"),
@@ -178,6 +201,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: ["G100"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -186,6 +210,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: ["G100"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2020"),
         record_qts_date: Date.parse("1/9/2017"),
@@ -194,6 +219,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["100403"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -202,6 +228,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["100403"],
         record_itt_date: Date.parse("1/9/2020"),
         record_qts_date: Date.parse("1/9/2017"),
@@ -210,6 +237,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: ["100403"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -218,14 +246,19 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: ["100403"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2020"),
         record_qts_date: Date.parse("1/9/2017"),
         record_qualification_name: "OTT"
       },
+      # end of mathematics for 2022
+
+      # start of physics for 2022
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: [],
+        record_itt_subjects: ["physics"],
         record_itt_subject_codes: ["F300"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -234,6 +267,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: [],
+        record_itt_subjects: ["physics"],
         record_itt_subject_codes: ["F300"],
         record_itt_date: Date.parse("1/9/2020"),
         record_qts_date: Date.parse("1/9/2017"),
@@ -242,6 +276,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: ["F300"],
+        record_itt_subjects: ["physics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -250,6 +285,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: ["F300"],
+        record_itt_subjects: ["physics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2020"),
         record_qts_date: Date.parse("1/9/2017"),
@@ -258,6 +294,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: [],
+        record_itt_subjects: ["physics"],
         record_itt_subject_codes: ["100425"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -266,6 +303,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: [],
+        record_itt_subjects: ["physics"],
         record_itt_subject_codes: ["100425"],
         record_itt_date: Date.parse("1/9/2020"),
         record_qts_date: Date.parse("1/9/2017"),
@@ -274,6 +312,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: ["100425"],
+        record_itt_subjects: ["physics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -282,14 +321,19 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: ["100425"],
+        record_itt_subjects: ["physics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2020"),
         record_qts_date: Date.parse("1/9/2017"),
         record_qualification_name: "Postgraduate Diploma in Education"
       },
+      # end of physics for 2022
+
+      # start of chemistry for 2022
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: [],
+        record_itt_subjects: ["chemistry"],
         record_itt_subject_codes: ["F100"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -298,6 +342,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: [],
+        record_itt_subjects: ["chemistry"],
         record_itt_subject_codes: ["F100"],
         record_itt_date: Date.parse("1/9/2020"),
         record_qts_date: Date.parse("1/9/2017"),
@@ -306,6 +351,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: ["F100"],
+        record_itt_subjects: ["chemistry"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -314,6 +360,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: ["F100"],
+        record_itt_subjects: ["chemistry"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2020"),
         record_qts_date: Date.parse("1/9/2017"),
@@ -322,6 +369,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: [],
+        record_itt_subjects: ["chemistry"],
         record_itt_subject_codes: ["100417"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -330,6 +378,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: [],
+        record_itt_subjects: ["chemistry"],
         record_itt_subject_codes: ["100417"],
         record_itt_date: Date.parse("1/9/2020"),
         record_qts_date: Date.parse("1/9/2017"),
@@ -338,6 +387,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: ["100417"],
+        record_itt_subjects: ["chemistry"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -346,14 +396,19 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: ["100417"],
+        record_itt_subjects: ["chemistry"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2020"),
         record_qts_date: Date.parse("1/9/2017"),
         record_qualification_name: "QTS Award only"
       },
+      # end of chemistry for 2022
+
+      # start of foreign_languages for 2022
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: [],
+        record_itt_subjects: ["foreign languages"],
         record_itt_subject_codes: ["Q100"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -362,6 +417,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: [],
+        record_itt_subjects: ["foreign languages"],
         record_itt_subject_codes: ["Q100"],
         record_itt_date: Date.parse("1/9/2020"),
         record_qts_date: Date.parse("1/9/2017"),
@@ -370,6 +426,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: ["Q100"],
+        record_itt_subjects: ["foreign languages"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -378,6 +435,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: ["Q100"],
+        record_itt_subjects: ["foreign languages"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2020"),
         record_qts_date: Date.parse("1/9/2017"),
@@ -386,6 +444,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: [],
+        record_itt_subjects: ["foreign languages"],
         record_itt_subject_codes: ["100321"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -394,6 +453,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: [],
+        record_itt_subjects: ["foreign languages"],
         record_itt_subject_codes: ["100321"],
         record_itt_date: Date.parse("1/9/2020"),
         record_qts_date: Date.parse("1/9/2017"),
@@ -402,6 +462,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: ["100321"],
+        record_itt_subjects: ["foreign languages"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -410,14 +471,19 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: ["100321"],
+        record_itt_subjects: ["foreign languages"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2020"),
         record_qts_date: Date.parse("1/9/2017"),
         record_qualification_name: "Teachers Certificate"
       },
+      # end of foreign_languages for 2022
+
+      # start of mathematics for 2023
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["G100"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2018"),
@@ -426,6 +492,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["G100"],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2017"),
@@ -434,6 +501,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: ["G100"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2018"),
@@ -442,6 +510,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["100403"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2018"),
@@ -450,6 +519,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: ["100403"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2018"),
@@ -458,6 +528,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["G100"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -466,6 +537,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: ["G100"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -474,6 +546,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["100403"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -482,14 +555,19 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: ["100403"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
         record_qualification_name: "BA"
       },
+      # end of mathematics for 2023
+
+      # start of physics for 2023
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: [],
+        record_itt_subjects: ["physics"],
         record_itt_subject_codes: ["F300"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -498,6 +576,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: ["F300"],
+        record_itt_subjects: ["physics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -506,6 +585,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: [],
+        record_itt_subjects: ["physics"],
         record_itt_subject_codes: ["100425"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -514,14 +594,19 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: ["100425"],
+        record_itt_subjects: ["physics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
         record_qualification_name: "BA"
       },
+      # end of physics for 2023
+
+      # start of chemistry for 2023
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: [],
+        record_itt_subjects: ["chemistry"],
         record_itt_subject_codes: ["F100"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -530,6 +615,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: ["F100"],
+        record_itt_subjects: ["chemistry"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -538,6 +624,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: [],
+        record_itt_subjects: ["chemistry"],
         record_itt_subject_codes: ["100417"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -546,14 +633,19 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: ["100417"],
+        record_itt_subjects: ["chemistry"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
         record_qualification_name: "BA"
       },
+      # end of chemistry for 2023
+
+      # start of foreign_languages for 2023
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: [],
+        record_itt_subjects: ["foreign languages"],
         record_itt_subject_codes: ["Q100"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -562,6 +654,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: ["Q100"],
+        record_itt_subjects: ["foreign languages"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -570,6 +663,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: [],
+        record_itt_subjects: ["foreign languages"],
         record_itt_subject_codes: ["100321"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -578,14 +672,19 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: ["100321"],
+        record_itt_subjects: ["foreign languages"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
         record_qualification_name: "BA"
       },
+      # end of foreign_languages for 2023
+
+      # start of mathematics for 2024
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["G100"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2019"),
@@ -594,6 +693,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: ["G100"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2019"),
@@ -602,6 +702,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["100403"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2019"),
@@ -610,6 +711,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: ["100403"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2019"),
@@ -618,6 +720,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["G100"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -626,6 +729,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: ["G100"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -634,6 +738,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["100403"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -642,14 +747,19 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: ["100403"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
         record_qualification_name: "BA"
       },
+      # end of mathematics for 2024
+
+      # start of physics for 2024
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: [],
+        record_itt_subjects: ["physics"],
         record_itt_subject_codes: ["F300"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -658,6 +768,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: ["F300"],
+        record_itt_subjects: ["physics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -666,6 +777,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: [],
+        record_itt_subjects: ["physics"],
         record_itt_subject_codes: ["100425"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -674,14 +786,19 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: ["100425"],
+        record_itt_subjects: ["physics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
         record_qualification_name: "BA"
       },
+      # end of physics for 2024
+
+      # start of chemistry for 2024
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: [],
+        record_itt_subjects: ["chemistry"],
         record_itt_subject_codes: ["F100"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -690,6 +807,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: ["F100"],
+        record_itt_subjects: ["chemistry"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -698,6 +816,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: [],
+        record_itt_subjects: ["chemistry"],
         record_itt_subject_codes: ["100417"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -706,14 +825,19 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: ["100417"],
+        record_itt_subjects: ["chemistry"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
         record_qualification_name: "BA"
       },
+      # end of chemistry for 2024
+
+      # start of foreign_languages for 2024
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: [],
+        record_itt_subjects: ["foreign languages"],
         record_itt_subject_codes: ["Q100"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -722,6 +846,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: ["Q100"],
+        record_itt_subjects: ["foreign languages"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -730,6 +855,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: [],
+        record_itt_subjects: ["foreign languages"],
         record_itt_subject_codes: ["100321"],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -738,11 +864,13 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: ["100321"],
+        record_itt_subjects: ["foreign languages"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2017"),
         record_qts_date: Date.parse("1/9/2020"),
         record_qualification_name: "BA"
       }
+      # end of foreign_languages for 2024
     ].each do |context|
       context "when claim academic year #{context[:claim_academic_year]}" do
         let(:claim_academic_year) { context[:claim_academic_year] }
@@ -750,19 +878,23 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
         context "when record degree codes #{context[:record_degree_codes]}" do
           let(:record_degree_codes) { context[:record_degree_codes] }
 
-          context "when record ITT subject codes #{context[:record_itt_subject_codes]}" do
-            let(:record_itt_subject_codes) { context[:record_itt_subject_codes] }
+          context "when record ITT subjects #{context[:record_itt_subjects]}" do
+            let(:record_itt_subjects) { context[:record_itt_subjects] }
 
-            context "when record QTS date #{context[:record_qts_date]}" do
-              let(:record_qts_date) { context[:record_qts_date] }
+            context "when record ITT subject codes #{context[:record_itt_subject_codes]}" do
+              let(:record_itt_subject_codes) { context[:record_itt_subject_codes] }
 
-              context "when record ITT date #{context[:record_itt_date]}" do
-                let(:record_itt_date) { context[:record_itt_date] }
+              context "when record QTS date #{context[:record_qts_date]}" do
+                let(:record_qts_date) { context[:record_qts_date] }
 
-                context "when record qualification name #{context[:record_qualification_name]}" do
-                  let(:record_qualification_name) { context[:record_qualification_name] }
+                context "when record ITT date #{context[:record_itt_date]}" do
+                  let(:record_itt_date) { context[:record_itt_date] }
 
-                  it { is_expected.to eql true }
+                  context "when record qualification name #{context[:record_qualification_name]}" do
+                    let(:record_qualification_name) { context[:record_qualification_name] }
+
+                    it { is_expected.to eql true }
+                  end
                 end
               end
             end
@@ -772,9 +904,11 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
     end
 
     [
+      # start of mathematics for 2021
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["G100"],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2019"),
@@ -783,6 +917,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: ["G100"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2019"),
@@ -791,6 +926,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["100403"],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2019"),
@@ -799,6 +935,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: ["100403"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2019"),
@@ -807,6 +944,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["G100"],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -815,6 +953,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: ["G100"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -823,6 +962,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["100403"],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -831,14 +971,19 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: ["100403"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2020"),
         record_qualification_name: "BA"
       },
+      # end of mathematics for 2021
+
+      # start of physics for 2021
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: [],
+        record_itt_subjects: ["physics"],
         record_itt_subject_codes: ["F300"],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -847,6 +992,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: ["F300"],
+        record_itt_subjects: ["physics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -855,6 +1001,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: [],
+        record_itt_subjects: ["physics"],
         record_itt_subject_codes: ["100425"],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -863,14 +1010,19 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: ["100425"],
+        record_itt_subjects: ["physics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2020"),
         record_qualification_name: "BA"
       },
+      # end of physics for 2021
+
+      # start of chemistry for 2021
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: [],
+        record_itt_subjects: ["Chemistry"],
         record_itt_subject_codes: ["F100"],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -879,6 +1031,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: ["F100"],
+        record_itt_subjects: ["Chemistry"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -887,6 +1040,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: [],
+        record_itt_subjects: ["Chemistry"],
         record_itt_subject_codes: ["100417"],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -895,6 +1049,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: ["100417"],
+        record_itt_subjects: ["Chemistry"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -903,14 +1058,19 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: [],
+        record_itt_subjects: ["Chemistry"],
         record_itt_subject_codes: ["Q100"],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2020"),
         record_qualification_name: "BA"
       },
+      # end of chemistry for 2021
+
+      # start of foreign_languages for 2021
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: ["Q100"],
+        record_itt_subjects: ["Foreign Languages"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -919,6 +1079,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: [],
+        record_itt_subjects: ["Foreign Languages"],
         record_itt_subject_codes: ["100321"],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2020"),
@@ -927,14 +1088,19 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2021),
         record_degree_codes: ["100321"],
+        record_itt_subjects: ["Foreign Languages"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2020"),
         record_qualification_name: "BA"
       },
+      # end of foreign_languages for 2021
+
+      # start of mathematics for 2022
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["G100"],
         record_itt_date: Date.parse("1/9/2019"),
         record_qts_date: Date.parse("1/9/2018"),
@@ -943,6 +1109,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: ["G100"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2019"),
         record_qts_date: Date.parse("1/9/2018"),
@@ -951,6 +1118,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["100403"],
         record_itt_date: Date.parse("1/9/2019"),
         record_qts_date: Date.parse("1/9/2018"),
@@ -959,14 +1127,19 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2022),
         record_degree_codes: ["100403"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2019"),
         record_qts_date: Date.parse("1/9/2018"),
         record_qualification_name: "BA"
       },
+      # end of mathematics for 2022
+
+      # start of mathematics for 2023
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["G100"],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2019"),
@@ -975,6 +1148,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: ["G100"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2019"),
@@ -983,6 +1157,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["100403"],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2019"),
@@ -991,14 +1166,19 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2023),
         record_degree_codes: ["100403"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2018"),
         record_qts_date: Date.parse("1/9/2019"),
         record_qualification_name: "BA"
       },
+      # end of mathematics for 2023
+
+      # start of mathematics for 2024
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["G100"],
         record_itt_date: Date.parse("1/9/2019"),
         record_qts_date: Date.parse("1/9/2018"),
@@ -1007,6 +1187,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: ["G100"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2019"),
         record_qts_date: Date.parse("1/9/2018"),
@@ -1015,6 +1196,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: [],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: ["100403"],
         record_itt_date: Date.parse("1/9/2019"),
         record_qts_date: Date.parse("1/9/2018"),
@@ -1023,6 +1205,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: ["100403"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2019"),
         record_qts_date: Date.parse("1/9/2018"),
@@ -1031,6 +1214,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2025),
         record_degree_codes: ["100403"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2019"),
         record_qts_date: Date.parse("1/9/2018"),
@@ -1039,6 +1223,7 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: ["1004030"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2019"),
         record_qts_date: Date.parse("1/9/2018"),
@@ -1047,11 +1232,13 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
       {
         claim_academic_year: AcademicYear.new(2024),
         record_degree_codes: ["100403"],
+        record_itt_subjects: ["mathematics"],
         record_itt_subject_codes: [],
         record_itt_date: Date.parse("1/9/2019"),
         record_qts_date: Date.parse("1/9/2010"),
         record_qualification_name: "BA"
       }
+      # end of mathematics for 2024
     ].each do |context|
       context "when claim academic year #{context[:claim_academic_year]}" do
         let(:claim_academic_year) { context[:claim_academic_year] }
@@ -1059,19 +1246,23 @@ RSpec.describe EarlyCareerPayments::DqtRecord do
         context "when record degree codes #{context[:record_degree_codes]}" do
           let(:record_degree_codes) { context[:record_degree_codes] }
 
-          context "when record ITT subject codes #{context[:record_itt_subject_codes]}" do
-            let(:record_itt_subject_codes) { context[:record_itt_subject_codes] }
+          context "when record ITT subjects #{context[:record_itt_subjects]}" do
+            let(:record_itt_subjects) { context[:record_itt_subjects] }
 
-            context "when record QTS date #{context[:record_qts_date]}" do
-              let(:record_qts_date) { context[:record_qts_date] }
+            context "when record ITT subject codes #{context[:record_itt_subject_codes]}" do
+              let(:record_itt_subject_codes) { context[:record_itt_subject_codes] }
 
-              context "when record ITT date #{context[:record_itt_date]}" do
-                let(:record_itt_date) { context[:record_itt_date] }
+              context "when record QTS date #{context[:record_qts_date]}" do
+                let(:record_qts_date) { context[:record_qts_date] }
 
-                context "when record qualification name #{context[:record_qualification_name]}" do
-                  let(:record_qualification_name) { context[:record_qualification_name] }
+                context "when record ITT date #{context[:record_itt_date]}" do
+                  let(:record_itt_date) { context[:record_itt_date] }
 
-                  it { is_expected.to eql false }
+                  context "when record qualification name #{context[:record_qualification_name]}" do
+                    let(:record_qualification_name) { context[:record_qualification_name] }
+
+                    it { is_expected.to eql false }
+                  end
                 end
               end
             end